edit : [협업] build를 위한 설정 및 버그 수정
- 해당 커밋은 강채원, 김혜민, 장주빈, 최나현이 함께 작성하였다.

- 최종 build를 위한 설정을 하였다.
- text 초기화가 안 되는 부분을 수정하였다.
- build시 문제가 되는 파일 업로드 부분을 수정하였다.


Former-commit-id: bf78a78b3f2bf8f0abe516a64fc715cb5bbd93e3

+++ epub_editor/src/background.js
@@ -35,9 +35,12 @@ ipcMain.on('upload', async (event, url, file, email, epubName, path) => {
     .then(function () {
       console.log('upload success')
       win.webContents.send('upload','success');
-      fs.unlink(file, (err) => {
-        console.log('널이니?',err)
-      })
+      let temp = file.split('.');
+      if (temp[temp.length-1] == 'zip') {
+        fs.unlink(file, (err) => {
+          console.log('널이니?',err)
+        })
+      }
     })
     .catch(function (err) {
       console.log(file, '에러에러에러')
@@ -63,26 +66,26 @@ async function createWindow() {
       webSecurity: false
     }
   });
-  const templete=[
-    {
-      label: 'edit',
-      submenu:[
-        {
-          role:'zoomIn',
-          accelerator: 'CommandOrControl+='
-        },
-        {
-          role:'zoomOut'
-        },
-        {
-          role:'reload'
-        }
-      ]
-    }
-  ];
-  let newMenu= Menu.buildFromTemplate(templete);
+  // const templete=[
+  //   {
+  //     label: 'edit',
+  //     submenu:[
+  //       {
+  //         role:'zoomIn',
+  //         accelerator: 'CommandOrControl+='
+  //       },
+  //       {
+  //         role:'zoomOut'
+  //       },
+  //       {
+  //         role:'reload'
+  //       }
+  //     ]
+  //   }
+  // ];
+  // let newMenu= Menu.buildFromTemplate(templete);
 
-  Menu.setApplicationMenu(newMenu);
+  // Menu.setApplicationMenu(newMenu);
 
   win.on('close', function(e) {
     //setTimeout(function () {console.log('isKor ' + isKor)}, 3000);
